refactor(nodes): split primary-node.ts into modular components (Issue #695)
Split the 1094-line primary-node.ts into smaller, focused modules:

| File | Lines | Description |
|------|-------|-------------|
| primary-node.ts | 709 | Core class (reduced from 1094, ~35% reduction) |
| local-execution-service.ts | 293 | AgentPool and local execution management |
| schedule-command-handler.ts | 176 | Schedule management commands |
| next-step-service.ts | 122 | Next-step recommendations |
| index.ts | 18 | Module exports |

## Key Changes

### LocalExecutionService
- AgentPool initialization and lifecycle
- SchedulerService initialization
- TaskFlowOrchestrator initialization
- Feedback channel management
- Local prompt/command execution

### ScheduleCommandHandler
- listSchedules, getSchedule, enableSchedule, disableSchedule, runSchedule

### NextStepService
- triggerNextStepRecommendation
- extractRecentMessages

## Benefits

- Better separation of concerns
- Easier testing of individual components
- Clearer responsibilities
- Improved maintainability

## Test Results

- 8 primary-node tests pass
- 1556 total tests pass (1 pre-existing LOG_LEVEL test failure unrelated)

Fixes #695
